refactor: Organize folder structure (#4)
